+++ internal/database/database.go
@@ -1,9 +1,11 @@
-package main
+package database
 
 import (
 	"database/sql"
 	"errors"
 	"fmt"
+	"github.com/atburke/teleport_interview/internal/types"
+	"github.com/atburke/teleport_interview/internal/crypto"
 	_ "github.com/go-sql-driver/mysql"
 	"time"
 )
@@ -11,46 +13,24 @@ import (
 // tokens and password hash could be []byte objects, but they're strings in the
 // db and in the api, so leaving them as hex strings will save time
 
-// Account represents a user's account and login information.
-type Account struct {
-	AccountId, Email, PasswordHash, Salt string
-}
-
-// Session represent's a user's session.
-type Session struct {
-	AccountId, SessionToken, CSRFToken string
-	ExpireIdle, ExpireAbs              time.Time
-}
-
-// Expired checks if the session has expired.
-func (session *Session) Expired(t time.Time) bool {
-	return t.After(session.ExpireIdle) || t.After(session.ExpireAbs)
-}
-
-// Authenticated returns true if the session is associated with an authenticated
-// user, and false if the session is pre-auth.
-func (session *Session) Authenticated() bool {
-	return session.AccountId != ""
-}
-
 // Database represents a database connection.
 type Database interface {
 
 	// CreatePreAuthSession creates a new session for a user that has not yet
 	// authenticated.
-	CreatePreAuthSession(csrfToken string, initTime time.Time) (*Session, error)
+	CreatePreAuthSession(csrfToken string, initTime time.Time) (*types.Session, error)
 
 	// CreateSession creates a new session for an authenticated user.
-	CreateSession(accountId, csrfToken string, initTime time.Time) (*Session, error)
+	CreateSession(accountId, csrfToken string, initTime time.Time) (*types.Session, error)
 
 	// FetchSession fetches the session identified by the provided token.
-	FetchSession(sessionToken string) (*Session, error)
+	FetchSession(sessionToken string) (*types.Session, error)
 
 	// DeleteSession deletes a session for a user.
 	DeleteSession(sessionToken string) error
 
 	// FetchAccount fetches the account for a user.
-	FetchAccount(email string) (*Account, error)
+	FetchAccount(email string) (*types.Account, error)
 
 	// DeleteExpiredSessions deletes sessions that have expired (after either their
 	// idle or absolute timeout has passed)
@@ -78,14 +58,14 @@ func NewMySqlDatabase(username, password, databaseName string) (*MySqlDatabase, 
 	return &MySqlDatabase{driver}, nil
 }
 
-func (db *MySqlDatabase) CreatePreAuthSession(csrfToken string, initTime time.Time) (*Session, error) {
+func (db *MySqlDatabase) CreatePreAuthSession(csrfToken string, initTime time.Time) (*types.Session, error) {
 	// an empty account ID is functionally equivalent to NULL. Consider doing this
 	// instead of having account_id nullable?
 	return db.CreateSession("", csrfToken, initTime)
 }
 
-func (db *MySqlDatabase) CreateSession(accountId, csrfToken string, initTime time.Time) (*Session, error) {
-	sessionToken, err := GenerateToken()
+func (db *MySqlDatabase) CreateSession(accountId, csrfToken string, initTime time.Time) (*types.Session, error) {
+	sessionToken, err := crypto.GenerateToken()
 	if err != nil {
 		return nil, fmt.Errorf("Cound not generate token: %w", err)
 	}
@@ -102,10 +82,10 @@ func (db *MySqlDatabase) CreateSession(accountId, csrfToken string, initTime tim
 		return nil, fmt.Errorf("Error inserting new session: %w", err)
 	}
 
-	return &Session{accountId, sessionToken, csrfToken, initTime, initTime}, nil
+	return &types.Session{accountId, sessionToken, csrfToken, initTime, initTime}, nil
 }
 
-func (db *MySqlDatabase) FetchSession(sessionToken string) (*Session, error) {
+func (db *MySqlDatabase) FetchSession(sessionToken string) (*types.Session, error) {
 	var accountIdRaw sql.NullString
 	var accountId, csrfToken string
 	var expireIdle, expireAbs time.Time
@@ -127,7 +107,7 @@ func (db *MySqlDatabase) FetchSession(sessionToken string) (*Session, error) {
 		accountId = accountIdRaw.String
 	}
 
-	return &Session{accountId, sessionToken, csrfToken, expireIdle, expireAbs}, nil
+	return &types.Session{accountId, sessionToken, csrfToken, expireIdle, expireAbs}, nil
 }
 
 func (db *MySqlDatabase) DeleteSession(sessionToken string) error {
@@ -139,7 +119,7 @@ func (db *MySqlDatabase) DeleteSession(sessionToken string) error {
 	return nil
 }
 
-func (db *MySqlDatabase) FetchAccount(email string) (*Account, error) {
+func (db *MySqlDatabase) FetchAccount(email string) (*types.Account, error) {
 	var accountId, passwordHash, salt string
 	stmt := "SELECT account_id, password_hash, salt FROM Accounts WHERE email = ?"
 	err := db.driver.QueryRow(
@@ -152,7 +132,7 @@ func (db *MySqlDatabase) FetchAccount(email string) (*Account, error) {
 		return nil, fmt.Errorf("SQL error: %w", err)
 	}
 
-	return &Account{accountId, email, passwordHash, salt}, nil
+	return &types.Account{accountId, email, passwordHash, salt}, nil
 }
 
 func (db *MySqlDatabase) DeleteExpiredSessions(t time.Time) error {
