+++ internal/database/database_test.go
@@ -1,0 +1,3 @@
+package database
+
+// tests in here would probably be just integration tests
