+++ internal/types/types_test.go
@@ -1,4 +1,4 @@
-package main
+package types
 
 import (
 	"github.com/stretchr/testify/assert"
